Interim fix
common.mk breaks when setting COMMON_NO_THIRD_PARTY_REQD, this can be
made better.

HDF5 headers are now downloaded on the fly when they aren't found in
HDF5_INC.  Again, this could be made better.  HDF requires you to run
autoconf in order to complete the header file set.  The headers
downloaded may not be portable.

We don't need for HDF libraries when compiling here, only when linking.
However, we'll need to make sure (somehow) that libraries conform to
the headers we're compiling against.

Compilation order matters here, but it shouldn't.